Moving server schema validation into schema file
