+++ lib/server.js
@@ -14,7 +14,7 @@ var Defaults = require('./defaults');
 var Boom = require('boom');
 var Request = require('./request');
 var Route = require('./route');
-var Validation = require('./validation');
+var Schema = require('./schema');
 var Views = require('./views');
 var Utils = require('./utils');
 // Pack delayed required inline
@@ -63,7 +63,7 @@ module.exports = internals.Server = function (/* host, port, options */) {      
     this.settings = Utils.applyToDefaults(Defaults.server, args.options || {});
 
     if (this.settings.strict === true) {
-        Validation.server(this.settings, function (err) {
+        Schema.server(this.settings, function (err) {
 
             Utils.assert(!err, 'Server options are invalid: ' + err);
         });
