+++ test/unit/validation.js
@@ -357,52 +357,4 @@ describe('Validation', function () {
             });
         });
     });
-
-    describe('#server', function () {
-
-        it('fails when unknown properties exist', function (done) {
-
-            var settings = { unknown: true };
-
-            Validation.server(settings, function (err) {
-
-                expect(err).to.exist;
-                done();
-            });
-        });
-
-        it('fails when unknown properties exist', function (done) {
-
-            var server = new Hapi.Server({ strict: false });
-            server.settings.unknown = true;
-
-            Validation.server(server.settings, function (err) {
-
-                expect(err).to.exist;
-                done();
-            });
-        });
-
-        it('succeeds with default settings returned from server', function (done) {
-
-            var server = new Hapi.Server();
-
-            Validation.server(server.settings, function (err) {
-
-                expect(err).to.not.exist;
-                done();
-            });
-        });
-
-        it('succeeds with only a couple of settings provided', function (done) {
-
-            var fn = function () {
-
-                var server = new Hapi.Server({ cache: 'memory', docs: true });
-            };
-
-            expect(fn).to.not.throw(Error);
-            done();
-        });
-    });
 });
